+++ NEWS.md
@@ -1,6 +1,8 @@
 rphylopic (development version)
 ==============
 
+* added text argument to get_attribution
+* get_attribution now handles multiple uuids
 * added functions for transforming PhyloPic silhouettes (flipping and rotating)
 * save_phylopic bg argument updated to be "transparent" by default
 * added geom_phylopic (#25)